backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Payments
stripe==7.8.0

# Validation & Config
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client for webhooks
httpx==0.26.0

